issue 46 fix ci tests

+++ internal/domain/requests/withdrawals.go
@@ -1,6 +1,6 @@
 package requests
 
 type Withdrawals struct {
-	Order string `json:"order"`
-	Sum   int64  `json:"sum"`
+	Order string  `json:"order"`
+	Sum   float64 `json:"sum"`
 }
